refactor directive
