+++ src/app/shared/components/fxdonad-shared/code-editor/code-editor.component.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/themes/variables";
+@use "../../../../styles/themes/variables";
 
 .code-editor-container {
   display: flex;
