--- src/app/shared/components/loading-overlay/loading-overlay.component.scss
@@ -1,38 +1,0 @@
-.loading-overlay {
-  position: absolute; /* hoặc fixed tuỳ ý che ở cấp nào */
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 9999;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  .spinner {
-    width: 48px;
-    height: 48px;
-    border: 5px solid #ccc;
-    border-top: 5px solid #fff;
-    border-radius: 50%;
-    animation: spin 1s linear infinite;
-    margin-bottom: 16px;
-  }
-
-  p {
-    font-size: 16px;
-    color: #fff;
-  }
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
